Added option to see years individualy
Added the posibility to see the years as an individual row. Simplifying the use for the users when a wide range of dates is available.

Tried to maintain the same asthetics of the calendar. To use simply pass the [showYears] parameter as true.

Also added comments to help others understand better the inner works of the package and help in its development.

Love the package, great work. Hope you like it and merge it into master.