Added preselection of first available date when month or year changes & improvements
Trying to enhance user experience, I updated the way dates are selected.

Currently a user must select the day last for the "selectedDate" to change. But I've seen that a lot of the times, the user selects the day first and then the month or year, and changes multiple times. And until he selects the day, the new date is not selected.
I believe that it is more intuitive that the state of the day or month stays the same, if posible, through the changes. So now:

-If a user selects a day, and then the month. In the new month the preselected day will be chosen.

EXAMPLE: User selects the 29th of jan. Then changes to March, the 29th of March is selected. In the case that the selected month doesn't have that date available(Feb, for example), it will select the first available date of the month. If no day is available, It will keep the previously selected date and fire the function "OnNoAvailableDate", so the developer can choose how to let the user Know that there are no available dates on that month and the last selected date is the current one.

I also added lints to encourage good coding practices and easier development.

Hope It helps as always!, Love this Package ;)